trading_system/requirements.txt
asyncio
aiohttp
websockets
numpy
pandas
pyyaml
python-dateutil
flask
requests
